Refactor ProfilePage and CompanyPage components for improved styling and functionality
- Updated ProfilePage.css to use CSS variables for colors and improved layout.
- Refactored ProfilePage.jsx to remove company management logic, focusing on user profile.
- Added CompanyPage.jsx for dedicated company management with CRUD functionality.
- Implemented responsive design for company cards and forms.
- Introduced dark mode toggle in CompanyPage.
- Enhanced user experience with success messages and improved form handling.
