+++ src/components/ErrorDisplay.jsx
@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from 'react';
 
-const ErrorDisplay = ({ error, onClose, duration = 5000 }) => {
+// Renamed to NotificationDisplay to match the updated context
+const NotificationDisplay = ({ notification, onClose, duration = 5000 }) => {
   const [visible, setVisible] = useState(true);
   
-  // Auto-hide the error after duration
+  // Auto-hide the notification after duration
   useEffect(() => {
-    if (!error) return;
+    if (!notification) return;
     
     const timer = setTimeout(() => {
       setVisible(false);
@@ -13,7 +14,7 @@ const ErrorDisplay = ({ error, onClose, duration = 5000 }) => {
     }, duration);
     
     return () => clearTimeout(timer);
-  }, [error, onClose, duration]);
+  }, [notification, onClose, duration]);
   
   // Handle manual close
   const handleClose = () => {
@@ -21,16 +22,20 @@ const ErrorDisplay = ({ error, onClose, duration = 5000 }) => {
     if (onClose) onClose();
   };
   
-  if (!error || !visible) return null;
+  if (!notification || !visible) return null;
+  
+  // Support for backward compatibility when notification is just a string
+  const message = typeof notification === 'string' ? notification : notification.message;
+  const type = typeof notification === 'string' ? 'error' : (notification.type || 'error');
   
   return (
-    <div className="error-display">
-      <div className="error-content">
-        <span className="error-message">{error}</span>
-        <button className="error-close" onClick={handleClose}>×</button>
+    <div className="notification-display">
+      <div className={`notification-content ${type}`}>
+        <span className="notification-message">{message}</span>
+        <button className="notification-close" onClick={handleClose}>×</button>
       </div>
     </div>
   );
 };
 
-export default ErrorDisplay;
+export default NotificationDisplay;
