+++ src/pages/ProfilePage.css
@@ -4,13 +4,13 @@
   max-width: 1200px;
   margin: 0 auto;
   padding: 2rem;
-  color: #333;
+  color: var(--text-color);
 }
 
 .profile-back-btn {
   display: inline-flex;
   align-items: center;
-  color: #666;
+  color: var(--light-text);
   text-decoration: none;
   margin-bottom: 2rem;
   transition: color 0.2s;
@@ -18,7 +18,7 @@
 }
 
 .profile-back-btn:hover {
-  color: #3498db;
+  color: var(--primary-color);
 }
 
 .profile-back-btn svg {
@@ -30,7 +30,7 @@
   align-items: center;
   margin-bottom: 2rem;
   padding-bottom: 1rem;
-  border-bottom: 1px solid #f0f0f0;
+  border-bottom: 1px solid var(--border-color);
 }
 
 .profile-avatar {
@@ -39,25 +39,25 @@
   border-radius: 50%;
   object-fit: cover;
   margin-right: 1.5rem;
-  border: 3px solid #f0f0f0;
+  border: 3px solid var(--border-color);
 }
 
 .profile-name {
   margin: 0 0 0.5rem 0;
   font-size: 1.8rem;
-  color: #333;
+  color: var(--text-color);
 }
 
 .profile-email {
   margin: 0;
-  color: #777;
+  color: var(--light-text);
   font-size: 1rem;
 }
 
 .profile-tabs {
   display: flex;
   margin-bottom: 2rem;
-  border-bottom: 1px solid #f0f0f0;
+  border-bottom: 1px solid var(--border-color);
 }
 
 .profile-tab {
@@ -67,7 +67,7 @@
   border-bottom: 2px solid transparent;
   margin-right: 1rem;
   font-size: 1rem;
-  color: #666;
+  color: var(--light-text);
   cursor: pointer;
   display: flex;
   align-items: center;
@@ -79,19 +79,19 @@
 }
 
 .profile-tab:hover {
-  color: #3498db;
+  color: var(--primary-color);
 }
 
 .profile-tab.active {
-  color: #3498db;
-  border-bottom: 2px solid #3498db;
+  color: var(--primary-color);
+  border-bottom: 2px solid var(--primary-color);
 }
 
 .profile-section {
-  background: #fff;
+  background: var(--card-bg);
   border-radius: 8px;
   padding: 1.5rem;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
+  box-shadow: var(--card-shadow);
   margin-bottom: 2rem;
 }
 
@@ -99,7 +99,7 @@
   margin-top: 0;
   margin-bottom: 1.5rem;
   font-weight: 600;
-  color: #333;
+  color: var(--text-color);
   font-size: 1.3rem;
 }
 
@@ -111,22 +111,24 @@
   display: block;
   margin-bottom: 0.5rem;
   font-weight: 500;
-  color: #555;
+  color: var(--light-text);
 }
 
 .form-group input,
 .form-group textarea {
   width: 100%;
   padding: 0.75rem;
-  border: 1px solid #ddd;
+  border: 1px solid var(--border-color);
   border-radius: 4px;
   font-size: 1rem;
   transition: border-color 0.2s;
+  background-color: var(--card-bg);
+  color: var(--text-color);
 }
 
 .form-group input:focus,
 .form-group textarea:focus {
-  border-color: #3498db;
+  border-color: var(--primary-color);
   outline: none;
   box-shadow: 0 0 0 2px rgba(52, 152, 219, 0.1);
 }
@@ -140,8 +142,7 @@
   flex: 1;
 }
 
-.profile-actions,
-.company-form-actions {
+.profile-actions {
   display: flex;
   justify-content: flex-end;
   gap: 1rem;
@@ -156,22 +157,22 @@
   font-weight: 500;
   cursor: pointer;
   transition: all 0.2s;
-  background-color: #3498db;
+  background-color: var(--primary-color);
   color: white;
 }
 
 .btn:hover {
-  background-color: #2980b9;
+  background-color: var(--primary-hover);
 }
 
 .btn:disabled {
-  background-color: #95a5a6;
+  background-color: var(--light-text);
   cursor: not-allowed;
 }
 
 .btn-secondary {
-  background-color: #f1f2f6;
-  color: #444;
+  background-color: var(--secondary-color);
+  color: var(--text-color);
   text-decoration: none;
   display: inline-flex;
   align-items: center;
@@ -179,7 +180,7 @@
 }
 
 .btn-secondary:hover {
-  background-color: #e6e7eb;
+  background-color: var(--border-color);
 }
 
 .section-header {
@@ -194,9 +195,9 @@
   align-items: center;
   padding: 0.5rem 1rem;
   background-color: transparent;
-  border: 1px solid #3498db;
+  border: 1px solid var(--primary-color);
   border-radius: 4px;
-  color: #3498db;
+  color: var(--primary-color);
   font-size: 0.9rem;
   cursor: pointer;
   transition: all 0.2s;
@@ -207,72 +208,8 @@
 }
 
 .btn-add:hover {
-  background-color: #3498db;
+  background-color: var(--primary-color);
   color: white;
-}
-
-.companies-list {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-  gap: 1.5rem;
-}
-
-.company-card {
-  border: 1px solid #f0f0f0;
-  border-radius: 8px;
-  overflow: hidden;
-  transition: all 0.2s;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
-}
-
-.company-card:hover {
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-  transform: translateY(-2px);
-}
-
-.company-card-header {
-  display: flex;
-  align-items: center;
-  padding: 1rem;
-  border-bottom: 1px solid #f0f0f0;
-}
-
-.company-logo {
-  width: 50px;
-  height: 50px;
-  border-radius: 4px;
-  object-fit: contain;
-  margin-right: 1rem;
-  background-color: #f9f9f9;
-}
-
-.company-info h4 {
-  margin: 0 0 0.25rem 0;
-  font-size: 1.1rem;
-}
-
-.company-gstin {
-  margin: 0;
-  font-size: 0.8rem;
-  color: #777;
-}
-
-.company-address {
-  padding: 1rem;
-  min-height: 60px;
-  color: #666;
-  font-size: 0.9rem;
-}
-
-.company-address p {
-  margin: 0;
-}
-
-.company-actions {
-  display: flex;
-  justify-content: flex-end;
-  padding: 0.75rem;
-  background-color: #f9f9f9;
 }
 
 .btn-icon {
@@ -283,109 +220,28 @@
   justify-content: center;
   border: none;
   border-radius: 4px;
-  background-color: #f1f2f6;
-  color: #555;
+  background-color: var(--secondary-color);
+  color: var(--light-text);
   margin-left: 0.5rem;
   cursor: pointer;
   transition: all 0.2s;
 }
 
 .btn-icon:hover {
-  background-color: #e6e7eb;
+  background-color: var(--border-color);
 }
 
 .btn-icon.danger {
-  color: #e74c3c;
+  color: var(--status-pending);
 }
 
 .btn-icon.danger:hover {
-  background-color: #fadbd8;
-}
-
-.no-companies {
-  text-align: center;
-  padding: 2rem;
-  color: #777;
-  background-color: #f9f9f9;
-  border-radius: 8px;
-}
-
-.no-companies p {
-  margin-bottom: 1rem;
-}
-
-.company-form {
-  background: #fff;
-  border-radius: 8px;
-  padding: 1.5rem;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
-}
-
-.company-form h3 {
-  margin-top: 0;
-  margin-bottom: 1.5rem;
-  font-weight: 600;
-  color: #333;
-  font-size: 1.3rem;
-}
-
-.logo-upload-section {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 2rem;
-}
-
-.company-logo-preview {
-  width: 120px;
-  height: 120px;
-  border-radius: 8px;
-  object-fit: contain;
-  margin-bottom: 1rem;
-  border: 1px solid #f0f0f0;
-  padding: 0.5rem;
-  background-color: #f9f9f9;
-}
-
-.btn-upload {
-  display: flex;
-  align-items: center;
-  padding: 0.5rem 1rem;
-  background-color: transparent;
-  border: 1px solid #3498db;
-  border-radius: 4px;
-  color: #3498db;
-  font-size: 0.9rem;
-  cursor: pointer;
-  transition: all 0.2s;
-  margin-bottom: 0.5rem;
-}
-
-.btn-upload svg {
-  margin-right: 0.5rem;
-}
-
-.btn-upload:hover {
-  background-color: #3498db;
-  color: white;
-}
-
-.company-bank-details {
-  margin-top: 2rem;
-  padding-top: 1.5rem;
-  border-top: 1px solid #f0f0f0;
-}
-
-.company-bank-details h4 {
-  margin-top: 0;
-  margin-bottom: 1.5rem;
-  color: #555;
-  font-size: 1.1rem;
+  background-color: rgba(231, 76, 60, 0.1);
 }
 
 .success-message {
-  background-color: #d4edda;
-  color: #155724;
+  background-color: rgba(16, 185, 129, 0.1);
+  color: var(--status-completed);
   padding: 1rem;
   border-radius: 4px;
   margin-bottom: 1.5rem;
@@ -404,17 +260,12 @@
     margin-bottom: 1rem;
   }
   
-  .companies-list {
-    grid-template-columns: 1fr;
-  }
-  
   .form-row {
     flex-direction: column;
     gap: 0;
   }
   
-  .profile-actions, 
-  .company-form-actions {
+  .profile-actions {
     flex-direction: column;
   }
